requirements.txt
asn1crypto==0.23.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
coloredlogs==7.3
cryptography==2.1.1
enum34==1.1.6
future==0.16.0
github3.py==1.0.0a4
humanfriendly==4.4.1
idna==2.6
ipaddress==1.0.18
monotonic==1.3
ndg-httpsclient==0.4.3
pgpdump==1.5
pyasn1==0.3.7
pycparser==2.18
pyOpenSSL==17.3.0
python-dateutil==2.6.1
requests==2.18.4
roca-detect==1.0.3
six==1.11.0
uritemplate==3.0.0
uritemplate.py==3.0.2
urllib3==1.22
